feat: 쿠폰 사용 가능 조건 커스텀 훅으로 변경
